chore(tree-cover-density): move widget from forest change to land cover The correct place to this widget is land cover according to the its ticket on JIRA
