+++ components/widgets/land-cover/tree-cover-density/selectors.js
