+++ components/widgets/manifest.js
@@ -16,7 +16,6 @@ import glads from 'components/widgets/forest-change/glads';
 // import gladRanked from 'components/widgets/forest-change/glad-ranked';
 import integratedAlertsRanked from 'components/widgets/forest-change/integrated-alerts-ranked';
 import integratedDeforestationAlerts from 'components/widgets/forest-change/integrated-deforestation-alerts';
-import treeCoverDensity from 'components/widgets/forest-change/tree-cover-density';
 
 // fires
 import firesAlerts from 'components/widgets/fires/fires-alerts';
@@ -42,6 +41,7 @@ import primaryForest from 'components/widgets/land-cover/primary-forest';
 import treeCoverLocated from 'components/widgets/land-cover/tree-cover-located';
 import USLandCover from 'components/widgets/land-cover/us-land-cover';
 import rankedForestTypes from 'components/widgets/land-cover/ranked-forest-types';
+import treeCoverDensity from 'components/widgets/land-cover/tree-cover-density';
 
 // Climate
 import woodyBiomass from 'components/widgets/climate/whrc-biomass/';
@@ -77,7 +77,6 @@ export default {
   integratedAlertsRanked,
   integratedDeforestationAlerts,
   netChange,
-  treeCoverDensity,
 
   // fires
   firesAlerts,
@@ -103,6 +102,7 @@ export default {
   primaryForest,
   treeCoverLocated,
   rankedForestTypes,
+  treeCoverDensity,
 
   // climate
   // emissions,
